+++ lib/connection/connection_config.js
@@ -13,8 +13,7 @@ const RowMode = require('./../constants/row_mode');
 const DataTypes = require('./result/data_types');
 const Logger = require('../logger');
 const { isSnowflakeHost } = require('../authentication/authentication_util');
-// TODO(SNOW-3548350): remove with the telemetry emission (target: 2026-11-30).
-const { recordInputShape } = require('./connection_identifier_shape');
+const { buildConnectionShapePayload } = require('../telemetry/connection_identifier_shape');
 const WAIT_FOR_BROWSER_ACTION_TIMEOUT = 120000;
 const SNOWFLAKE_AUTHORIZE_ENDPOINT = '/oauth/authorize';
 const SNOWFLAKE_TOKEN_REQUEST_ENDPOINT = '/oauth/token-request';
@@ -87,25 +86,6 @@ const DEFAULT_PARAMS = [
 ];
 
 function consolidateHostAndAccount(options) {
-  // TODO(SNOW-3548350): remove this capture with the
-  // `client_connection_identifier_shape` telemetry emission
-  // (target: 2026-11-30).
-  //
-  // Capture the user-supplied connection-identifier shape BEFORE any
-  // normalization runs (the dotted-account split below mutates
-  // `options.account`, the `-org` strip near the bottom does too, and
-  // the `Util.constructHostname` synthesis fills in `options.host` /
-  // `options.accessUrl` for the bare-account case). This is the only
-  // entry point both the programmatic-options path and the
-  // `connections.toml` path go through, so a single capture site here
-  // covers both.
-  //
-  // Returned to the caller (the `ConnectionConfig` constructor) rather
-  // than couriered through `options._inputShape`, so we do not mutate
-  // the public options object with a temporary internal key that has to
-  // be cleaned up afterwards.
-  const inputShape = recordInputShape(options);
-
   let dotPos = -1;
   let realAccount = undefined;
   let realRegion = undefined;
@@ -176,13 +156,6 @@ function consolidateHostAndAccount(options) {
   }
   options.account = realAccount;
   options.region = realRegion;
-
-  // TODO(SNOW-3548350): remove with the telemetry emission
-  // (target: 2026-11-30). Returned to the caller (the `ConnectionConfig`
-  // constructor) so the shape can be stored on the instance without
-  // having to leak a temporary `_inputShape` key onto the public options
-  // object.
-  return inputShape;
 }
 
 /**
@@ -207,6 +180,9 @@ function ConnectionConfig(options, validateCredentials, qaMode, clientInfo) {
 
   // check for invalid options
   Errors.checkArgumentValid(Util.isObject(options), ErrorCodes.ERR_CONN_CREATE_INVALID_OPTIONS);
+
+  // TODO(SNOW-3548350): remove with the in-band telemetry emission (target: 2026-11-30).
+  this.connectionShapeForTelemetry = buildConnectionShapePayload(options);
 
   // only validate credentials if necessary
   if (validateCredentials) {
@@ -336,16 +312,7 @@ function ConnectionConfig(options, validateCredentials, qaMode, clientInfo) {
     }
   }
 
-  // TODO(SNOW-3548350): remove with the telemetry emission
-  // (target: 2026-11-30). `consolidateHostAndAccount` captures the
-  // user-supplied connection-identifier shape before normalizing the
-  // `options` object and hands it back as its return value, so we can
-  // stash it on this instance without leaking a temporary internal key
-  // onto the public options object or touching the public typings
-  // (`WIP_ConnectionOptions`, `index.d.ts`). Consumed by `_getInputShape`
-  // at the `client_connection_identifier_shape` emit site in
-  // `lib/services/sf.js`.
-  this._inputShape = consolidateHostAndAccount(options);
+  consolidateHostAndAccount(options);
 
   // check for missing account
   Errors.checkArgumentExists(
@@ -1368,23 +1335,6 @@ ConnectionConfig.prototype.getRetrySfStartingSleepTime = function () {
 
 ConnectionConfig.prototype.getRetrySfMaxSleepTime = function () {
   return this._getParameterValue(PARAM_RETRY_SF_MAX_SLEEP_TIME);
-};
-
-/**
- * Returns the user-supplied connection-identifier shape captured by
- * `consolidateHostAndAccount`, or undefined when the constructor was
- * called via a path that bypasses shape capture (most notably the
- * `deserializeConnection` path). Internal — consumed only by the
- * `client_connection_identifier_shape` in-band telemetry emit site.
- *
- * TODO(SNOW-3548350): remove with the telemetry emission
- * (target: 2026-11-30).
- *
- * @returns {import('./connection_identifier_shape').ConnectionIdentifierShape|undefined}
- * @private
- */
-ConnectionConfig.prototype._getInputShape = function () {
-  return this._inputShape;
 };
 
 /**
