+++ lib/services/sf.js
@@ -62,13 +62,7 @@ const sfParams = require('../constants/sf_params');
 const { getMinicoreStatus } = require('../minicore');
 const { getOsDetails } = require('../telemetry/os_details');
 const { getApplicationPath } = require('../telemetry/application_path');
-const {
-  buildInbandTelemetryRequest,
-  CLIENT_CONNECTION_IDENTIFIER_SHAPE,
-  isConnectionShapeTelemetryDisabledByEnv,
-} = require('../telemetry/inband_telemetry');
-// TODO(SNOW-3548350): remove with the telemetry emission (target: 2026-11-30).
-const { shapeToTelemetryPayload } = require('../connection/connection_identifier_shape');
+const { buildInbandTelemetryRequest } = require('../telemetry/inband_telemetry');
 const { getDetectedPlatforms } = require('../telemetry/platform_detection');
 const { getLibcDetails } = require('../telemetry/libc_details');
 const { getSpcsToken } = require('../authentication/spcs_token');
@@ -1248,37 +1242,19 @@ StateConnecting.prototype.continue = async function () {
           });
       }
 
-      // TODO(SNOW-3548350): remove this block when the
-      // connection-identifier-shape data collection wraps up
-      // (target: 2026-11-30).
-      //
-      // Emit one in-band telemetry record per successful login describing
-      // which connection-identifier fields the user supplied (no hostname
-      // or account value is included). Mirrors PR snowflakedb/gosnowflake#1797
-      // (Go) and PR snowflakedb/snowflake-connector-python#2877 (Python).
-      // Like the `minicore_error` emit above this is unconditional —
-      // it does not gate on `CLIENT_TELEMETRY_ENABLED`, matching the
-      // existing in-band-telemetry precedent in this driver and
-      // intentionally diverging from Go/Python/JDBC. The local kill
-      // switch is `SF_TELEMETRY_DISABLE_CONNECTION_SHAPE=true`.
-      if (!isConnectionShapeTelemetryDisabledByEnv()) {
-        const inputShape =
-          typeof connectionConfig._getInputShape === 'function'
-            ? connectionConfig._getInputShape()
-            : undefined;
-        if (inputShape) {
-          const shapeRequest = buildInbandTelemetryRequest(
-            connectionConfig,
-            CLIENT_CONNECTION_IDENTIFIER_SHAPE,
-            shapeToTelemetryPayload(inputShape),
-          );
-          parent
-            .createSessionTokenRequest(shapeRequest)
-            .sendAsync()
-            .catch((error) => {
-              Logger.getInstance().debug(`Failed to send telemetry event: %s`, error);
-            });
-        }
+      // TODO(SNOW-3548350): remove with the in-band telemetry emission (target: 2026-11-30).
+      if (!process.env.SF_TELEMETRY_DISABLE_CONNECTION_SHAPE) {
+        const shapeRequest = buildInbandTelemetryRequest(
+          connectionConfig,
+          'client_connection_identifier_shape',
+          connectionConfig.connectionShapeForTelemetry,
+        );
+        parent
+          .createSessionTokenRequest(shapeRequest)
+          .sendAsync()
+          .catch((error) => {
+            Logger.getInstance().debug(`Failed to send telemetry event: %s`, error);
+          });
       }
     } else {
       if (Errors.isNetworkError(err) || Errors.isRequestFailedError(err)) {
